requirements.txt
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
flask_cors
flask
python-dotenv
pytesseract
fastapi
pydantic
passlib
python-jose
uvicorn